Finished the reorganization of tests
