+++ Tests/Fixtures/AbstractTestCase.php
@@ -45,16 +45,6 @@ abstract class AbstractTestCase extends WebTestCase
      *
      * @return Crawler
      */
-    protected function doGetRequest(array $parameters = array())
-    {
-        return $this->client->request('GET', '/admin/'.(empty($parameters) ? '' : '?'.http_build_query($parameters, '', '&')));
-    }
-
-    /**
-     * @param array $parameters
-     *
-     * @return Crawler
-     */
     protected function getBackendPage(array $parameters)
     {
         return $this->client->request('GET', '/admin/?'.http_build_query($parameters, '', '&'));
@@ -67,6 +57,6 @@ abstract class AbstractTestCase extends WebTestCase
      */
     protected function getBackendHomepage()
     {
-        return $this->doGetRequest(array('entity' => 'Category', 'view' => 'list'));
+        return $this->getBackendPage(array('entity' => 'Category', 'view' => 'list'));
     }
 }
