+++ Tests/bootstrap.php
@@ -44,7 +44,7 @@ if (is_dir($buildDir = __DIR__.'/../build')) {
 
 include __DIR__.'/Fixtures/App/AppKernel.php';
 
-$application = new Application(new AppKernel('test', true));
+$application = new Application(new AppKernel('default_backend', true));
 $application->setAutoExit(false);
 
 // Create database
